simplify API and move to own folder
