+++ python_modules/libraries/dagster-elt/dagster_elt/__init__.py
@@ -1,17 +1,4 @@
 from dagster._core.libraries import DagsterLibraryRegistry
-from .resources import SlingResource, SlingSourceConfig, SlingMode, SlingTarget
-from .asset_defs import build_sling_asset
-
-
 from .version import __version__
 
 DagsterLibraryRegistry.register("dagster-elt", __version__)
-
-__all__ = [
-    "SlingResource",
-    "SlingSourceConfig",
-    "SlingMode",
-    "SlingTarget",
-    "SlingSyncConfig",
-    "build_sling_asset",
-]
